Kill LocationTestBase and clean up imports
