+++ corehq/apps/locations/tests/test_location_groups.py
@@ -1,17 +1,38 @@
 from mock import patch
-from corehq.apps.locations.fixtures import location_fixture_generator
-from corehq.apps.locations.tests.util import make_loc
-from corehq.apps.locations.tests.test_locations import LocationTestBase
+
+from django.test import TestCase
+
+from corehq.apps.commtrack.tests.util import bootstrap_location_types
+from corehq.apps.domain.shortcuts import create_domain
 from corehq.apps.groups.exceptions import CantSaveException
 from corehq.apps.users.models import CommCareUser
 from corehq.util.test_utils import flag_enabled
 
+from ..fixtures import location_fixture_generator
+from .util import make_loc
+
 
 @flag_enabled('HIERARCHICAL_LOCATION_FIXTURE')
-class LocationGroupTest(LocationTestBase):
+class LocationGroupTest(TestCase):
 
     def setUp(self):
         super(LocationGroupTest, self).setUp()
+        self.domain = create_domain('locations-test')
+        self.domain.convert_to_commtrack()
+        bootstrap_location_types(self.domain.name)
+
+        self.loc = make_loc('loc', type='outlet', domain=self.domain.name)
+        self.sp = self.loc.linked_supply_point()
+
+        self.user = CommCareUser.create(
+            self.domain.name,
+            'username',
+            'password',
+            first_name='Bob',
+            last_name='Builder',
+        )
+        self.user.set_location(self.loc)
+
         self.test_state = make_loc(
             'teststate',
             type='state',
@@ -29,6 +50,10 @@ class LocationGroupTest(LocationTestBase):
             parent=self.test_village,
             domain=self.domain.name
         )
+
+    def tearDown(self):
+        self.domain.delete()
+        super(LocationGroupTest, self).tearDown()
 
     def test_group_name(self):
         # just location name for top level
